Remove RoleManagerTests and enhance UserTests
The `RoleManagerTests.cs` file has been removed, eliminating tests for role management functionality. The `UserTests.cs` file has been updated to include new tests for admin promotions and demotions, event queuing on role changes, and restrictions for non-admin users. Additional checks for invalid role changes have also been added.
